Initial commit for bioinfcourse
This contains the source files for my course in Bioinformatics and
Genomics started in the autumn of 2015.

The files are mostly latex with a bit of perl and R (and some tikz).
The most interesting / useful thing here are probably the perl scripts
that implement Needleman-Wunsch and Smith-Waterman and output a
visualisation for latex/beamer presentations using tikz. That's
quite neat, but more to the point guarantees the correct values
get put into the figures (I was prompted to do this as I found
mistakes in published books and could not imagine I would be
any less likely to make similar mistakes).